requirements.txt
beautifulsoup4>=4.10.0
curlify>=2.2.1
python-dateutil
pytz>=2021.3
certifi>=2024.6.2
geopy >= 2.2.0
cloudscraper>=1.2.60

# For tests
python-dotenv==1.1.0
ruff==0.11.10
ipython==9.2.0
pytest==8.3.5
pytest-mock==3.14.1
pytest-clarity==1.0.1
pytest-socket==0.7.0